requirements.txt
aniso8601==10.0.0
# apturl==0.5.2
bcrypt==3.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
binaryfile==1.2.0
blinker==1.9.0
# Brlapi==0.8.3
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
# command-not-found==0.3
# configobj==5.0.6
cryptography==3.4.8
# cupshelpers==1.0
dalybms==0.5.0
dbfread==2.0.4
# dbus-python==1.2.18
# defer==1.0.6
distro==1.7.0
# distro-info==1.1+ubuntu0.2
# DmxPy==0.5
# duplicity==0.8.21
fasteners==0.14.1
Flask==3.1.0
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
future==0.18.2
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
httplib2==0.20.2
idna==3.3
ijson==2.6.1
importlib-metadata==4.6.4
iso8601==2.1.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
# kazam==1.4.5
keyring==23.5.0
# language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
# louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
# meza==0.46.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.1.2
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==2.2.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
# PyAudio==0.2.14
# pycairo==1.20.1
# pycups==2.0.1
# PyGObject==3.42.1
pygogo==1.3.0
PyJWT==2.3.0
pylibftdi==0.23.0
pymacaroons==0.13.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
# python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
# python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.0
python-engineio==4.11.2
python-slugify==1.2.6
python-socketio==5.12.1
pytz==2022.1
pyxdg==0.27
# PyYAML==5.4.1
reportlab==3.6.8
requests==2.32.3
# screen-resolution-extra==0.0.0
SecretStorage==3.3.1
simple-websocket==1.1.0
six==1.16.0
soupsieve==2.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.37
ssh-import-id==5.11
# systemd-python==234
# terminator==2.1.1
# typing_extensions==4.12.2
tzdata==2024.2
# ubuntu-drivers-common==0.0.0
# ubuntu-pro-client==8001
# ufw==0.36.1
# unattended-upgrades==0.1
Unidecode==1.3.8
urllib3==1.26.5
# usb-creator==0.3.7
# wadllib==1.3.6
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
